+++ tests/endpoints/test_preview.py
@@ -1,5 +1,6 @@
 from fastapi.testclient import TestClient
 from pytest_mock.plugin import MockerFixture
+from pytubefix.exceptions import VideoPrivate, LoginRequired
 
 from ytdl_api.schemas.responses import VideoInfoResponse
 
@@ -39,8 +40,26 @@ def test_get_preview_422(app_client: TestClient):
     assert json_response["detail"][0]["msg"] == "Bad youtube video link provided."
 
 
-def test_get_preview_private_video(app_client: TestClient):
+def test_get_preview_private_video(
+    app_client: TestClient,
+    mocker: MockerFixture,
+):
+    mocker.patch("ytdl_api.downloaders.PytubeDownloader.get_video_info", side_effect=VideoPrivate("mCk1ChMlqt0"))
     response = app_client.get("/api/preview", params={"url": "https://www.youtube.com/watch?v=mCk1ChMlqt0"})
     assert response.status_code == 403
     json_response = response.json()
     assert "is a private video" in json_response["detail"]
+
+
+def test_get_preview_login_required(
+    app_client: TestClient,
+    mocker: MockerFixture,
+):
+    mocker.patch(
+        "ytdl_api.downloaders.PytubeDownloader.get_video_info",
+        side_effect=LoginRequired("mCk1ChMlqt0", "login required"),
+    )
+    response = app_client.get("/api/preview", params={"url": "https://www.youtube.com/watch?v=mCk1ChMlqt0"})
+    assert response.status_code == 403
+    json_response = response.json()
+    assert "login required" in json_response["detail"]