+++ ytdl_api/config.py
@@ -91,6 +91,10 @@ class Settings(BaseConfig):
 
     @property
     def downloader_version(self) -> str:
+        if self.downloader == DownloaderType.PYTUBE:
+            return version("pytubefix")
+        if self.downloader == DownloaderType.MOCK:
+            return "mock"
         return version(self.downloader.value)
 
     @validator("remove_expired_downloads_task_cron", pre=True)